Craps Game
Rolls two dice and sums them. If the sum is 2, 3, or 12, you lose; if
the sum is 7 or 11, you win; if the sum is another value, a point is
established. In that event, it continues rolling until either a 7 or
the same point value is rolled. If 7 is rolled, you lose. Otherwise,
you win.